Combine `ModCounterDisplay` and `DifficultyMultiplierDisplay` into one class
